Rename abstract-group to abstract-group.component
See https://angular.io/styleguide.
