+++ src/client/app/humidity-sensor/humidity-sensors-group.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
-import { AbstractGroupComponent } from '../shared/abstract-groups-items/abstract-group';
+import { AbstractGroupComponent } from '../shared/abstract-groups-items/abstract-group.component';
 import { HumiditySensorViewModel } from './humidity-sensor-view-model';
 
 /**.
