+++ src/client/app/local-episodic-effect/local-episodic-effects-group.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
-import { AbstractGroupComponent } from '../shared/abstract-groups-items/abstract-group';
+import { AbstractGroupComponent } from '../shared/abstract-groups-items/abstract-group.component';
 import { LocalEpisodicEffectViewModel } from './local-episodic-effect-view-model';
 
 /**.
