+++ src/client/app/water-vapor-sensor/water-vapor-sensors-group.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
 import { FormBuilder } from '@angular/forms';
-import { AbstractGroupComponent } from '../shared/abstract-groups-items/abstract-group';
+import { AbstractGroupComponent } from '../shared/abstract-groups-items/abstract-group.component';
 import { WaterVaporSensorViewModel } from './water-vapor-sensor-view-model';
 
 /**
